fix: alterando as variáveis comparacoes e movimentacoes para long em InsertionSort

+++ src/InsertionSort.java
@@ -9,12 +9,12 @@ public class InsertionSort<T extends Comparable<T>> implements IOrdenador<T> {
     private double nanoToMilli = 1.0/1_000_000;
 
     @Override
-    public int getComparacoes() {
+    public long getComparacoes() {
         return comparacoes;
     }
 
     @Override
-    public int getMovimentacoes() {
+    public long getMovimentacoes() {
         return movimentacoes;
     }
